backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
anthropic
requests
